Convert gui progress
The commit to github hasnt really been working properly so this commit will be considered as the 3rd commit.
Sorry for the inconsistencY
Changes, updates, and disccusion
Removed To_farenheit and To_Celcius functions due to repetitive code
The converter_Gui is mostly completed and the conversion calculations, rounding, and overall formatting work well
Had to restart the converter gui 03 as it was easier to start from v2 than trouble shooting a difficult error i had encountered.
Removed degree_symbl code as its dumb and too much code to insert a single character